fix: resolve blocking and important scaffolding issues
Frontend:
- Fix API paths to match backend endpoints (/trends/metrics)
- Update response types to use paginated format
- Add missing start_date_local field to Activity interface

Backend models:
- Change ID columns to BigInteger (prevent overflow)
- Change distance/speed/elevation to Numeric (precision)
- Change JSON columns to JSONB (queryable)
- Add unique constraints (user_id+strava_id, activity+stream_type, etc.)
- Add composite indexes for query performance

Backend API:
- Add start_date_local to activity responses
- Add nested computed_metrics and effort_groups to activity detail

Alembic:
- Initialize Alembic configuration
- Create initial migration with all tables, constraints, and indexes

+++ backend/models/route.py
@@ -1,33 +1,35 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey, Text
+from sqlalchemy import Column, String, DateTime, ForeignKey, Index, Text
+from sqlalchemy.dialects.postgresql import BIGINT, NUMERIC
 from sqlalchemy.sql import func
 from database import Base
+
 
 class Route(Base):
     __tablename__ = "routes"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BIGINT, primary_key=True, index=True)
     name = Column(String(255))
-    
+
     # GPS data
     polyline = Column(Text)  # encoded polyline
     polyline_hash = Column(String(64), unique=True, nullable=False, index=True)
-    
+
     # Metrics
-    distance = Column(Float)  # meters
-    elevation_gain = Column(Float)  # meters
-    
+    distance = Column(NUMERIC(10, 2))  # meters
+    elevation_gain = Column(NUMERIC(10, 2))  # meters
+
     # Start/end coordinates
-    start_lat = Column(Float)
-    start_lng = Column(Float)
-    end_lat = Column(Float)
-    end_lng = Column(Float)
-    
+    start_lat = Column(NUMERIC(9, 6))
+    start_lng = Column(NUMERIC(9, 6))
+    end_lat = Column(NUMERIC(9, 6))
+    end_lng = Column(NUMERIC(9, 6))
+
     # Clustering
-    cluster_id = Column(Integer, ForeignKey("route_clusters.id", ondelete="SET NULL"), index=True)
-    
+    cluster_id = Column(BIGINT, ForeignKey("route_clusters.id", ondelete="SET NULL"), index=True)
+
     # Activity count (denormalized)
-    activity_count = Column(Integer, default=0)
-    
+    activity_count = Column(BIGINT, default=0)
+
     # Metadata
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
